% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octostep.R
\name{any.null}
\alias{all.null}
\alias{any.null}
\title{Check multiple arguments for \code{NULL}}
\usage{
\method{any}{null}(...)

\method{all}{null}(...)
}
\arguments{
\item{...}{Any R object.}
}
\value{
Logical.
}
\description{
Check multiple arguments for \code{NULL}
}

