% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octostep.R
\name{octostep}
\alias{octostep}
\title{Iterate lists getting a \emph{window} argument list to your callback.}
\usage{
octostep(x, func, pad = 1L, use.names = TRUE, transform.previous = FALSE)
}
\arguments{
\item{x}{List \strong{required}.}

\item{func}{Function with \code{length(formals(func))} == 
\code{2L * pad + 1L} \strong{required}.}

\item{pad}{Integer controlling how many items should be padded \emph{around} 
the current item, must be within \code{1L:((length(x) - 1L) / 2L)} for the 
ordinary use case with \code{length(x) >= 2L * pad + 1L}, otherwise 1L 
\strong{optional}.}

\item{use.names}{If \code{x} has a \code{names} attribute, should it be 
used for the return value? \strong{optional}.}

\item{transform.previous}{Should the \code{previous} arguments of the 
callback take the values of previous callbacks rather than the plain
values of the initial input list \strong{optional}.}
}
\value{
List.
}
\description{
Besides a list \code{octostep} takes a function as input, 
applies it to each element of the list and returns the resulting list.
}
\seealso{
\code{\link{mergeLists}}
}

